requirements.txt
torch>=2.0.0
absl-py
numpy
matplotlib